feat: mpsc log

+++ src/arch/aarch64/interrupt.cpp
@@ -12,10 +12,9 @@
 namespace {
 auto DefaultInterruptHandler(uint64_t cause, cpu_io::TrapContext* context)
     -> uint64_t {
-  klog::Info(
-      "Default Interrupt handler 0x%llX, 0x%llx",
-      static_cast<unsigned long long>(cause),
-      static_cast<unsigned long long>(reinterpret_cast<uintptr_t>(context)));
+  klog::Info("Default Interrupt handler 0x%llX, 0x%llx",
+             static_cast<uint64_t>(cause),
+             static_cast<uint64_t>(reinterpret_cast<uintptr_t>(context)));
   return 0;
 }
 }  // namespace
