+++ src/arch/riscv64/interrupt.cpp
@@ -17,21 +17,19 @@
 namespace {
 auto DefaultInterruptHandler(uint64_t cause, cpu_io::TrapContext* context)
     -> uint64_t {
-  klog::Info(
-      "Default Interrupt handler [%s] 0x%llX, 0x%llx",
-      cpu_io::ScauseInfo::kInterruptNames[cause],
-      static_cast<unsigned long long>(cause),
-      static_cast<unsigned long long>(reinterpret_cast<uintptr_t>(context)));
+  klog::Info("Default Interrupt handler [%s] 0x%llX, 0x%llx",
+             cpu_io::ScauseInfo::kInterruptNames[cause],
+             static_cast<uint64_t>(cause),
+             static_cast<uint64_t>(reinterpret_cast<uintptr_t>(context)));
   return 0;
 }
 
 auto DefaultExceptionHandler(uint64_t cause, cpu_io::TrapContext* context)
     -> uint64_t {
-  klog::Err(
-      "Default Exception handler [%s] 0x%llX, 0x%llx",
-      cpu_io::ScauseInfo::kExceptionNames[cause],
-      static_cast<unsigned long long>(cause),
-      static_cast<unsigned long long>(reinterpret_cast<uintptr_t>(context)));
+  klog::Err("Default Exception handler [%s] 0x%llX, 0x%llx",
+            cpu_io::ScauseInfo::kExceptionNames[cause],
+            static_cast<uint64_t>(cause),
+            static_cast<uint64_t>(reinterpret_cast<uintptr_t>(context)));
   while (true) {
     cpu_io::Pause();
   }
@@ -78,7 +76,7 @@ void Interrupt::RegisterInterruptFunc(uint64_t cause, InterruptDelegate func) {
     interrupt_handlers_[exception_code] = func;
     klog::Info("RegisterInterruptFunc [%s] 0x%llX",
                cpu_io::ScauseInfo::kInterruptNames[exception_code],
-               static_cast<unsigned long long>(cause));
+               static_cast<uint64_t>(cause));
   } else {
     assert(exception_code < cpu_io::ScauseInfo::kExceptionMaxCount &&
            "Exception code out of range");
@@ -86,7 +84,7 @@ void Interrupt::RegisterInterruptFunc(uint64_t cause, InterruptDelegate func) {
     exception_handlers_[exception_code] = func;
     klog::Info("RegisterInterruptFunc [%s] 0x%llX",
                cpu_io::ScauseInfo::kExceptionNames[exception_code],
-               static_cast<unsigned long long>(cause));
+               static_cast<uint64_t>(cause));
   }
 }
