+++ src/arch/riscv64/plic/plic.cpp
@@ -13,10 +13,9 @@
 namespace {
 auto DefaultPlicHandler(uint64_t cause, cpu_io::TrapContext* context)
     -> uint64_t {
-  klog::Info(
-      "Default PLIC handler, cause: 0x%llx, context: 0x%llx",
-      static_cast<unsigned long long>(cause),
-      static_cast<unsigned long long>(reinterpret_cast<uintptr_t>(context)));
+  klog::Info("Default PLIC handler, cause: 0x%llx, context: 0x%llx",
+             static_cast<uint64_t>(cause),
+             static_cast<uint64_t>(reinterpret_cast<uintptr_t>(context)));
   return 0;
 }
 }  // namespace
