+++ src/filesystem/fatfs/diskio.cpp
@@ -64,9 +64,8 @@ DRESULT disk_read(BYTE pdrv, BYTE* buff, LBA_t sector, UINT count) {
   }
   auto result = dev->ReadSectors(sector, static_cast<uint32_t>(count), buff);
   if (!result) {
-    klog::Err("disk_read: pdrv=%u sector=%llu count=%u failed",
-              static_cast<unsigned>(pdrv),
-              static_cast<unsigned long long>(sector),
+    klog::Err("disk_read: pdrv=%u sector=%lu count=%u failed",
+              static_cast<unsigned>(pdrv), static_cast<uint64_t>(sector),
               static_cast<unsigned>(count));
     return RES_ERROR;
   }
@@ -89,9 +88,8 @@ DRESULT disk_write(BYTE pdrv, const BYTE* buff, LBA_t sector, UINT count) {
   }
   auto result = dev->WriteSectors(sector, static_cast<uint32_t>(count), buff);
   if (!result) {
-    klog::Err("disk_write: pdrv=%u sector=%llu count=%u failed",
-              static_cast<unsigned>(pdrv),
-              static_cast<unsigned long long>(sector),
+    klog::Err("disk_write: pdrv=%u sector=%lu count=%u failed",
+              static_cast<unsigned>(pdrv), static_cast<uint64_t>(sector),
               static_cast<unsigned>(count));
     return RES_ERROR;
   }
