+++ src/main.cpp
@@ -34,7 +34,7 @@ std::atomic<uint64_t> global_counter{0};
 /// Task1: 每 1s 打印一次，测试 sys_exit
 void task1_func(void* arg) {
   klog::Info("Task1: arg = 0x%llx",
-             static_cast<unsigned long long>(reinterpret_cast<uintptr_t>(arg)));
+             static_cast<uint64_t>(reinterpret_cast<uintptr_t>(arg)));
   for (int i = 0; i < 5; ++i) {
     klog::Info("Task1: iteration %d/5", i + 1);
     sys_sleep(1000);
@@ -46,11 +46,10 @@ void task1_func(void* arg) {
 /// Task2: 每 2s 打印一次，测试 sys_yield
 void task2_func(void* arg) {
   klog::Info("Task2: arg = 0x%llx",
-             static_cast<unsigned long long>(reinterpret_cast<uintptr_t>(arg)));
+             static_cast<uint64_t>(reinterpret_cast<uintptr_t>(arg)));
   uint64_t count = 0;
   while (1) {
-    klog::Info("Task2: yield count=%llu",
-               static_cast<unsigned long long>(count++));
+    klog::Info("Task2: yield count=%lu", static_cast<uint64_t>(count++));
     sys_sleep(2000);
     // 主动让出 CPU
     sys_yield();
@@ -60,12 +59,12 @@ void task2_func(void* arg) {
 /// Task3: 每 3s 打印一次，同时修改全局变量，测试多核同步
 void task3_func(void* arg) {
   klog::Info("Task3: arg = 0x%llx",
-             static_cast<unsigned long long>(reinterpret_cast<uintptr_t>(arg)));
+             static_cast<uint64_t>(reinterpret_cast<uintptr_t>(arg)));
   while (1) {
     uint64_t old_value = global_counter.fetch_add(1, std::memory_order_seq_cst);
-    klog::Info("Task3: global_counter %llu -> %llu",
-               static_cast<unsigned long long>(old_value),
-               static_cast<unsigned long long>(old_value + 1));
+    klog::Info("Task3: global_counter %lu -> %lu",
+               static_cast<uint64_t>(old_value),
+               static_cast<uint64_t>(old_value + 1));
     sys_sleep(3000);
   }
 }
@@ -73,17 +72,17 @@ void task3_func(void* arg) {
 /// Task4: 每 4s 打印一次，测试 sys_sleep
 void task4_func(void* arg) {
   klog::Info("Task4: arg = 0x%llx",
-             static_cast<unsigned long long>(reinterpret_cast<uintptr_t>(arg)));
+             static_cast<uint64_t>(reinterpret_cast<uintptr_t>(arg)));
   uint64_t iteration = 0;
   while (1) {
     auto* cpu_sched = per_cpu::GetCurrentCore().sched_data;
     auto start_tick = cpu_sched->local_tick;
-    klog::Info("Task4: sleeping for 4s (iteration %llu)",
-               static_cast<unsigned long long>(iteration++));
+    klog::Info("Task4: sleeping for 4s (iteration %lu)",
+               static_cast<uint64_t>(iteration++));
     sys_sleep(4000);
     auto end_tick = cpu_sched->local_tick;
-    klog::Info("Task4: woke up (slept ~%llu ticks)",
-               static_cast<unsigned long long>(end_tick - start_tick));
+    klog::Info("Task4: woke up (slept ~%lu ticks)",
+               static_cast<uint64_t>(end_tick - start_tick));
   }
 }
